Rename plugins folder
