+++ src/formElementPlugins/dropdownChoice/src/SummaryView.tsx
