+++ src/formElementPlugins/textInfo/pluginConfig.json
