+++ src/formElementPlugins/textInfo/src/ApplicationView.tsx
